6996: Properly support converters
