+++ agent/src/main/java/org/openjdk/jmc/agent/Agent.java
@@ -57,7 +57,7 @@ public class Agent {
 	/**
 	 * This should be generated as part of the build later.
 	 */
-	public final static String VERSION = "0.0.2"; //$NON-NLS-1$
+	public final static String VERSION = "0.1.0"; //$NON-NLS-1$
 	private static boolean loadedDynamically = false;
 
 	/**
