+++ agent/src/main/java/org/openjdk/jmc/agent/Transformer.java
@@ -46,8 +46,7 @@ import org.openjdk.jmc.agent.jfr.JFRTransformDescriptor;
 import org.openjdk.jmc.agent.jfr.VersionResolver;
 import org.openjdk.jmc.agent.jfr.VersionResolver.JFRVersion;
 import org.openjdk.jmc.agent.jfr.impl.JFRClassVisitor;
-import org.openjdk.jmc.agent.jfrnext.impl.JFRNextClassVisitor;
-import org.openjdk.jmc.agent.util.InspectionClassLoader;
+import org.openjdk.jmc.agent.jfrlegacy.impl.JFRLegacyClassVisitor;
 
 public class Transformer implements ClassFileTransformer {
 	private TransformRegistry registry;
@@ -61,19 +60,20 @@ public class Transformer implements ClassFileTransformer {
 		ClassLoader loader, String className, Class<?> classBeingRedefined, ProtectionDomain protectionDomain,
 		byte[] classfileBuffer) throws IllegalClassFormatException {
 
-		// We need a class instance for reflective inspection, so create a InspectionClassLoader if the class if not yet 
-		// loaded.
 		return doTransforms(registry.getTransformData(className), classfileBuffer, loader, classBeingRedefined,
-				protectionDomain, classBeingRedefined != null ? null : new InspectionClassLoader(loader));
+				protectionDomain);
 	}
 
 	private byte[] doTransforms(
 		List<TransformDescriptor> transformDataList, byte[] classfileBuffer, ClassLoader definingClassLoader,
-		Class<?> classBeingRedefined, ProtectionDomain protectionDomain, InspectionClassLoader inspectionClassLoader) {
+		Class<?> classBeingRedefined, ProtectionDomain protectionDomain) {
+		if (transformDataList == null) {
+			return null;
+		}
 		for (TransformDescriptor td : transformDataList) {
 			// FIXME: Optimization, should do all transforms to one class in one go, instead of creating one class writer per transform.
 			classfileBuffer = doTransform(td, classfileBuffer, definingClassLoader, classBeingRedefined,
-					protectionDomain, inspectionClassLoader);
+					protectionDomain);
 			td.setPendingTransforms(false);
 		}
 		return classfileBuffer;
@@ -81,14 +81,14 @@ public class Transformer implements ClassFileTransformer {
 
 	private byte[] doTransform(
 		TransformDescriptor td, byte[] classfileBuffer, ClassLoader definingClassLoader, Class<?> classBeingRedefined,
-		ProtectionDomain protectionDomain, InspectionClassLoader inspectionClassLoader) {
+		ProtectionDomain protectionDomain) {
 		return doJFRLogging((JFRTransformDescriptor) td, classfileBuffer, definingClassLoader, classBeingRedefined,
-				protectionDomain, inspectionClassLoader);
+				protectionDomain);
 	}
 
 	private byte[] doJFRLogging(
 		JFRTransformDescriptor td, byte[] classfileBuffer, ClassLoader definingClassLoader,
-		Class<?> classBeingRedefined, ProtectionDomain protectionDomain, InspectionClassLoader inspectionClassLoader) {
+		Class<?> classBeingRedefined, ProtectionDomain protectionDomain) {
 		if (VersionResolver.getAvailableJFRVersion() == JFRVersion.NONE) {
 			Logger.getLogger(getClass().getName()).log(Level.SEVERE,
 					"Could not find JFR classes. Failed to instrument " + td.getMethod().toString()); //$NON-NLS-1$
@@ -97,10 +97,9 @@ public class Transformer implements ClassFileTransformer {
 		try {
 			ClassWriter classWriter = new ClassWriter(ClassWriter.COMPUTE_MAXS);
 			ClassVisitor visitor = VersionResolver.getAvailableJFRVersion() == JFRVersion.JFRNEXT
-					? new JFRNextClassVisitor(classWriter, td, definingClassLoader, classBeingRedefined,
-							protectionDomain, inspectionClassLoader)
-					: new JFRClassVisitor(classWriter, td, definingClassLoader, classBeingRedefined, protectionDomain,
-							inspectionClassLoader);
+					? new JFRClassVisitor(classWriter, td, definingClassLoader, classBeingRedefined, protectionDomain)
+					: new JFRLegacyClassVisitor(classWriter, td, definingClassLoader, classBeingRedefined,
+							protectionDomain);
 			ClassReader reader = new ClassReader(classfileBuffer);
 			reader.accept(visitor, ClassReader.EXPAND_FRAMES);
 			return classWriter.toByteArray();
