+++ agent/src/main/java/org/openjdk/jmc/agent/converters/FileConverter.java
@@ -1,6 +1,7 @@
 /*
- * Copyright (c) 2018, Oracle and/or its affiliates. All rights reserved.
- * 
+ * Copyright (c) 2020, Oracle and/or its affiliates. All rights reserved.
+ * Copyright (c) 2020, Datadog, Inc. All rights reserved.
+ *
  * DO NOT ALTER OR REMOVE COPYRIGHT NOTICES OR THIS FILE HEADER.
  *
  * The contents of this file are subject to the terms of either the Universal Permissive License
@@ -32,16 +33,23 @@
  */
 package org.openjdk.jmc.agent.converters;
 
+import java.io.File;
+
+import org.openjdk.jmc.agent.Agent;
+
 /**
- * Interface for converting objects to Strings.
+ * A converter which converts files into their canonical path.
  */
-public interface StringConverter<T> {
-	/**
-	 * Converts an object to a String.
-	 * 
-	 * @param o
-	 *            the object to convert.
-	 * @return the object converted to a String.
-	 */
-	String convert(T o);
+public final class FileConverter {
+	public static String convert(File file) {
+		if (file == null) {
+			return "null";
+		}
+		try {
+			return file.getCanonicalPath();
+		} catch (Throwable e) {
+			Agent.getLogger().warning("Agent failed to convert file to String: " + e.getMessage());
+		}
+		return file.getPath();
+	}
 }
