+++ agent/src/main/java/org/openjdk/jmc/agent/jfr/impl/JFREventClassGenerator.java
@@ -41,44 +41,35 @@ import org.objectweb.asm.Label;
 import org.objectweb.asm.MethodVisitor;
 import org.objectweb.asm.Opcodes;
 import org.objectweb.asm.Type;
+import org.openjdk.jmc.agent.Agent;
 import org.openjdk.jmc.agent.Attribute;
-import org.openjdk.jmc.agent.Field;
 import org.openjdk.jmc.agent.Parameter;
+import org.openjdk.jmc.agent.Field;
 import org.openjdk.jmc.agent.ReturnValue;
+import org.openjdk.jmc.agent.impl.ResolvedConvertable;
 import org.openjdk.jmc.agent.jfr.JFRTransformDescriptor;
 import org.openjdk.jmc.agent.util.TypeUtils;
 import org.openjdk.jmc.agent.util.expression.IllegalSyntaxException;
 
+/**
+ * This class is responsible for generating the JFR event class described by the transform
+ * descriptor.
+ */
 public class JFREventClassGenerator {
-	private static final String CLASS_NAME_INSTANT_EVENT = "com/oracle/jrockit/jfr/InstantEvent"; //$NON-NLS-1$
-	private static final String CLASS_NAME_DURATION_EVENT = "com/oracle/jrockit/jfr/DurationEvent"; //$NON-NLS-1$
-	private static final String CLASS_NAME_TIMED_EVENT = "com/oracle/jrockit/jfr/TimedEvent"; //$NON-NLS-1$
+	private static final String CLASS_EVENT = "jdk/jfr/Event"; //$NON-NLS-1$
 
-	/**
-	 * Generates an event class.
-	 * 
-	 * @param td
-	 *            the transform descriptor describing the transform.
-	 * @return returns the byte code for the generated class.
-	 * @throws Exception
-	 *             if the event class could not be generated.
-	 */
 	public static byte[] generateEventClass(JFRTransformDescriptor td, Class<?> classBeingRedefined) throws Exception {
 		ClassWriter cw = new ClassWriter(0);
-		// TODO: Add support for different locations
-		cw.visit(Opcodes.V1_7, Opcodes.ACC_PUBLIC + Opcodes.ACC_SUPER, td.getEventClassName(), null,
-				getEventTypeName(JFREventType.TIMED), null);
+		// FIXME: Perhaps switch to Opcodes V9 when there is one.
+		cw.visit(Opcodes.V1_8, Opcodes.ACC_PUBLIC + Opcodes.ACC_SUPER, td.getEventClassName(), null, CLASS_EVENT, null);
 
 		cw.visitSource(TypeUtils.getNamePart(td.getEventClassName()) + TypeUtils.JAVA_FILE_EXTENSION, null);
 
 		String parameterizedClassName = TypeUtils.parameterize(td.getEventClassName());
 		generateClassAnnotations(cw, td);
-		generateTokenField(cw);
 		generateAttributeFields(cw, td, classBeingRedefined);
-		generateClinit(cw, td.getEventClassName(), parameterizedClassName);
 		generateInit(cw, td.getEventClassName(), parameterizedClassName);
 		cw.visitEnd();
-
 		return cw.toByteArray();
 	}
 
@@ -98,36 +89,54 @@ public class JFREventClassGenerator {
 	}
 
 	private static void createField(ClassWriter cw, JFRTransformDescriptor td, Attribute attribute, Type type) {
-		if (!td.isAllowedFieldType(type)) {
+		if (!td.isAllowedEventFieldType(attribute, type)) {
 			Logger.getLogger(JFREventClassGenerator.class.getName())
-					.warning("Skipped generating field in event class for parameter " + attribute + " and type " + type //$NON-NLS-1$ //$NON-NLS-2$
+					.warning("Skipped generating field in event class for attribute " + attribute + " and type " + type //$NON-NLS-1$ //$NON-NLS-2$
 							+ " because of configuration settings!"); //$NON-NLS-1$
 			return;
 		}
 
-		String fieldType = getFieldType(type);
+		String fieldType = null;
 
-		FieldVisitor fv = cw.visitField(Opcodes.ACC_PUBLIC, attribute.getFieldName(), fieldType, null, null);
-		AnnotationVisitor av = fv.visitAnnotation("Lcom/oracle/jrockit/jfr/ValueDefinition;", true); //$NON-NLS-1$
-		if (attribute.getName() != null) {
-			av.visit("name", attribute.getName()); //$NON-NLS-1$
+		if (attribute.hasConverter()) {
+			ResolvedConvertable resolved = new ResolvedConvertable(attribute.getConverterClassName());
+			fieldType = getFieldType(Type.getType(resolved.getConverterMethod().getReturnType()));
+		} else {
+			fieldType = getFieldType(type);
 		}
-		if (attribute.getDescription() != null) {
-			av.visit("description", attribute.getDescription()); //$NON-NLS-1$
-		}
-		if (attribute.getContentType() != null) {
-			av.visitEnum("contentType", "Lcom/oracle/jrockit/jfr/ContentType;", attribute.getContentType()); //$NON-NLS-1$ //$NON-NLS-2$
-		}
-		if (attribute.getRelationKey() != null) {
-			av.visit("relationKey", attribute.getRelationKey()); //$NON-NLS-1$
-		}
+
+		FieldVisitor fv = cw.visitField(Opcodes.ACC_PROTECTED, attribute.getFieldName(), fieldType, null, null);
+
+		// Name
+		AnnotationVisitor av = fv.visitAnnotation("Ljdk/jfr/Label;", true);
+		av.visit("value", attribute.getName());
 		av.visitEnd();
+
+		// Description
+		av = fv.visitAnnotation("Ljdk/jfr/Description;", true);
+		av.visit("value", attribute.getDescription());
+		av.visitEnd();
+
+		// "ContentType"
+		// We support the old JDK 7 style content types transparently.
+		// We also support user defined content types and a single string value annotation parameter to the annotation.
+		String contentTypeAnnotation = getContentTypeAnnotation(attribute.getContentType());
+		if (contentTypeAnnotation != null) {
+			String[] contentTypeAnnotationInfo = contentTypeAnnotation.split(";");
+			av = fv.visitAnnotation(contentTypeAnnotationInfo[0] + ";", true);
+			if (contentTypeAnnotationInfo.length > 1) {
+				av.visit("value", contentTypeAnnotationInfo[1]);
+			}
+			av.visitEnd();
+		}
+
+		// FIXME: RelKey
 		fv.visitEnd();
 	}
 
 	private static void createField(ClassWriter cw, JFRTransformDescriptor td, Type type) {
 		ReturnValue returnValue = td.getReturnValue();
-		if (!td.isAllowedFieldType(type)) {
+		if (!td.isAllowedEventFieldType(returnValue, type)) {
 			Logger.getLogger(JFREventClassGenerator.class.getName())
 					.warning("Skipped generating field in event class for return value " + returnValue + " and type " //$NON-NLS-1$//$NON-NLS-2$
 							+ type + " because of configuration settings!"); //$NON-NLS-1$
@@ -136,89 +145,124 @@ public class JFREventClassGenerator {
 
 		String fieldType = getFieldType(type);
 
-		FieldVisitor fv = cw.visitField(Opcodes.ACC_PUBLIC, returnValue.getFieldName(), fieldType, null, null);
-		AnnotationVisitor av = fv.visitAnnotation("Lcom/oracle/jrockit/jfr/ValueDefinition;", true); //$NON-NLS-1$
-		if (returnValue.getName() != null) {
-			av.visit("name", returnValue.getName()); //$NON-NLS-1$
-		}
-		if (returnValue.getDescription() != null) {
-			av.visit("description", returnValue.getDescription()); //$NON-NLS-1$
-		}
-		if (returnValue.getContentType() != null) {
-			av.visitEnum("contentType", "Lcom/oracle/jrockit/jfr/ContentType;", returnValue.getContentType()); //$NON-NLS-1$ //$NON-NLS-2$
-		}
+		FieldVisitor fv = cw.visitField(Opcodes.ACC_PROTECTED, returnValue.getFieldName(), fieldType, null, null);
+
+		// Name
+		AnnotationVisitor av = fv.visitAnnotation("Ljdk/jfr/Label;", true);
+		av.visit("value", returnValue.getName());
 		av.visitEnd();
+
+		// Description
+		av = fv.visitAnnotation("Ljdk/jfr/Description;", true);
+		av.visit("value", returnValue.getDescription());
+		av.visitEnd();
+
+		// "ContentType"
+		// We support the old JDK 7 style content types transparently.
+		// We also support user defined content types and a single string value annotation parameter to the annotation.
+		String contentTypeAnnotation = getContentTypeAnnotation(returnValue.getContentType());
+		if (contentTypeAnnotation != null) {
+			String[] contentTypeAnnotationInfo = contentTypeAnnotation.split(";");
+			av = fv.visitAnnotation(contentTypeAnnotationInfo[0] + ";", true);
+			if (contentTypeAnnotationInfo.length > 1) {
+				av.visit("value", contentTypeAnnotationInfo[1]);
+			}
+			av.visitEnd();
+		}
+
+		// FIXME: RelKey
 		fv.visitEnd();
+	}
+
+	private static String getContentTypeAnnotation(String contentType) {
+		if (contentType == null) {
+			return null;
+		}
+		switch (contentType) {
+		case "None":
+			return null;
+		case "Address":
+			return "Ljdk/jfr/MemoryAddress;";
+		case "Bytes":
+			return "Ljdk/jfr/DataAmount;";
+		case "Timestamp":
+			return "Ljdk/jfr/Timestamp;";
+		case "Millis":
+			return "Ljdk/jfr/Timespan;" + "MILLISECONDS";
+		case "Nanos":
+			return "Ljdk/jfr/Timespan;" + "NANOSECONDS";
+		case "Ticks":
+			return "Ljdk/jfr/Timespan;" + "TICKS";
+		case "Percentage":
+			return "Ljdk/jfr/Percentage;";
+
+		default:
+			if (contentType.startsWith("L") && contentType.endsWith(";")) {
+				Agent.getLogger()
+						.fine("Using user defined content type. Note that this only works with JDK 9 and later!");
+				return contentType;
+			}
+			Agent.getLogger().severe("Unsupported content type " + contentType
+					+ ". Either use a JDK 7/8 content type, or specify the class of the annotation specifying the content type, e.g. Ljdk/jfr/DataAmount;. If specifying the content type annotation explicitly, it will only work on JDK 9 or later.");
+			return null;
+		}
 	}
 
 	private static String getFieldType(Type type) {
 		if (type.getSort() == Type.OBJECT || type.getSort() == Type.ARRAY) {
 			return "Ljava/lang/String;"; //$NON-NLS-1$
 		}
-
 		return type.getDescriptor();
 	}
 
+	/*
+	 * In JDK 9 the event types do not really need to be registered. There are also no tokens to
+	 * track to optimize lookup. So no need for a clinit.
+	 *
+	 * That said, once the class has been defined, we will still register it, to make sure that the
+	 * metadata is visible to all consumers, even though no event has been emitted.
+	 */
 	private static void generateInit(ClassWriter cw, String className, String parameterizedClassName) {
-		MethodVisitor mv = cw.visitMethod(Opcodes.ACC_PUBLIC, "<init>", "()V", null, null); //$NON-NLS-1$ //$NON-NLS-2$
+		MethodVisitor mv = cw.visitMethod(Opcodes.ACC_PUBLIC, "<init>", "()V", null, null);
 		mv.visitCode();
 		Label l0 = new Label();
 		mv.visitLabel(l0);
-		mv.visitLineNumber(19, l0);
 		mv.visitVarInsn(Opcodes.ALOAD, 0);
-		mv.visitFieldInsn(Opcodes.GETSTATIC, className, "token", "Lcom/oracle/jrockit/jfr/EventToken;"); //$NON-NLS-1$ //$NON-NLS-2$
-		mv.visitMethodInsn(Opcodes.INVOKESPECIAL, CLASS_NAME_TIMED_EVENT, "<init>", //$NON-NLS-1$
-				"(Lcom/oracle/jrockit/jfr/EventToken;)V", false); //$NON-NLS-1$
+		mv.visitMethodInsn(Opcodes.INVOKESPECIAL, "jdk/jfr/Event", "<init>", "()V", false);
+		mv.visitInsn(Opcodes.RETURN);
 		Label l1 = new Label();
 		mv.visitLabel(l1);
-		mv.visitLineNumber(20, l1);
-		mv.visitInsn(Opcodes.RETURN);
-		Label l2 = new Label();
-		mv.visitLabel(l2);
-		mv.visitLocalVariable("this", parameterizedClassName, null, l0, l2, 0); //$NON-NLS-1$
-		mv.visitMaxs(2, 1);
+		mv.visitLocalVariable("this", parameterizedClassName, null, l0, l1, 0);
+		mv.visitMaxs(1, 1);
 		mv.visitEnd();
-	}
-
-	private static void generateClinit(ClassWriter cw, String className, String parameterizedClassName) {
-		MethodVisitor mv = cw.visitMethod(Opcodes.ACC_STATIC, "<clinit>", "()V", null, null); //$NON-NLS-1$ //$NON-NLS-2$
-		mv.visitCode();
-		mv.visitLdcInsn(Type.getType(parameterizedClassName));
-		mv.visitMethodInsn(Opcodes.INVOKESTATIC, JFRUtils.INAME, "register", //$NON-NLS-1$
-				"(Ljava/lang/Class;)Ljava/lang/Object;", false); //$NON-NLS-1$
-		mv.visitTypeInsn(Opcodes.CHECKCAST, "com/oracle/jrockit/jfr/EventToken");
-		mv.visitFieldInsn(Opcodes.PUTSTATIC, className, "token", "Lcom/oracle/jrockit/jfr/EventToken;"); //$NON-NLS-1$ //$NON-NLS-2$
-		mv.visitInsn(Opcodes.RETURN);
-		mv.visitMaxs(1, 0);
-		mv.visitEnd();
-	}
-
-	private static void generateTokenField(ClassWriter cw) {
-		FieldVisitor fv = cw.visitField(Opcodes.ACC_FINAL + Opcodes.ACC_STATIC, "token", //$NON-NLS-1$
-				"Lcom/oracle/jrockit/jfr/EventToken;", null, null); //$NON-NLS-1$
-		fv.visitEnd();
 	}
 
 	private static void generateClassAnnotations(ClassWriter cw, JFRTransformDescriptor td) {
-		AnnotationVisitor av0 = cw.visitAnnotation("Lcom/oracle/jrockit/jfr/EventDefinition;", true); //$NON-NLS-1$
-		av0.visit("name", td.getEventName()); //$NON-NLS-1$
-		av0.visit("description", td.getEventDescription()); //$NON-NLS-1$
-		av0.visit("path", td.getEventPath()); //$NON-NLS-1$
-		av0.visit("stacktrace", td.isRecordStackTrace()); //$NON-NLS-1$
-		av0.visit("thread", true); //$NON-NLS-1$
-		av0.visitEnd();
-	}
+		// Label
+		AnnotationVisitor av = cw.visitAnnotation("Ljdk/jfr/Label;", true);
+		av.visit("value", td.getEventName());
+		av.visitEnd();
 
-	private static String getEventTypeName(JFREventType eventType) {
-		switch (eventType) {
-		case DURATION:
-			return CLASS_NAME_DURATION_EVENT;
-		case TIMED:
-			return CLASS_NAME_TIMED_EVENT;
-		case INSTANT:
-			return CLASS_NAME_INSTANT_EVENT;
-		case UNDEFINED:
+		// Description
+		av = cw.visitAnnotation("Ljdk/jfr/Description;", true);
+		av.visit("value", td.getEventDescription());
+		av.visitEnd();
+
+		// Category (path)
+		String[] pathElements = td.getEventPath().split("/");
+		av = cw.visitAnnotation("Ljdk/jfr/Category;", true);
+		AnnotationVisitor arrayVisitor = av.visitArray("value");
+		for (String pathElement : pathElements) {
+			arrayVisitor.visit(null, pathElement);
 		}
-		return CLASS_NAME_TIMED_EVENT;
+		arrayVisitor.visitEnd();
+		av.visitEnd();
+
+		// Stacktrace on/off
+		av = cw.visitAnnotation("Ljdk/jfr/StackTrace;", true);
+		av.visit("value", td.isRecordStackTrace());
+		av.visitEnd();
+
+		// Note that thread is always recorded these days
 	}
 }
