+++ agent/src/main/resources/org/openjdk/jmc/agent/impl/jfrprobes_schema.xsd
@@ -36,10 +36,17 @@
 					</xs:documentation>
 				</xs:annotation>
 			</xs:element>
+			<xs:element type="xs:boolean" name="enabledefaultconverters" minOccurs="0">
+				<xs:annotation>
+					<xs:documentation>
+						Enables a set of pre-packaged converters, converting some commonly found types to Strings.
+					</xs:documentation>
+				</xs:annotation>
+			</xs:element>
 			<xs:element type="xs:boolean" name="emitonexception" minOccurs="0">
 				<xs:annotation>
 					<xs:documentation>
-						Causes events to only be emitted if the instrmented method throws an exception
+						Causes events to only be emitted if the instrumented method throws an exception
 					</xs:documentation>
 				</xs:annotation>
 			</xs:element>
