+++ configuration/ide/eclipse/dictionary/dictionary.txt
@@ -70,3 +70,4 @@ wildcard
 wildcards
 zip
 zipped
+datadog
